// packages/types/src/pet/enums.ts
export enum PetType {
	FOX = 'fox',
	CAT = 'cat',
	DOG = 'dog',
	BUNNY = 'bunny',
	CHICKEN = 'chicken',
	PANDA = 'panda',
	TURTLE = 'turtle',
	PARROT = 'parrot',
}

export enum PetNeed {
	HUNGER = 'hunger',
	THIRST = 'thirst',
	ENERGY = 'energy',
	HYGIENE = 'hygiene',
	TOILET = 'toilet',
}

export enum PetActionCategory {
	FEED = 'Feed',
	DRINK = 'Drink',
	WASH = 'Wash',
	CARE = 'Care',
}

export enum PetAnimation {
	EAT = 'eat',
	DRINK = 'drink',
	TOILET = 'toilet',
	BATH = 'bath',
	SLEEP = 'sleep',
}